requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0

# Code Quality
black>=24.0,<25.0
flake8>=7.0,<8.0
isort>=5.13,<6.0
pre-commit>=3.6,<4.0

# Development tools
django-debug-toolbar>=4.2,<5.0
django-extensions>=3.2,<4.0
ipython>=8.20,<9.0

# Type checking
mypy>=1.8,<2.0
django-stubs>=4.2,<5.0
